Make sure we join ALL threads before saving the df for the last time
Sometimes the thread might have obtained the last lead but hasn't finished
yet when the while loop makes its check. We want the thread to collect its
data and put it in the dataframe before we save.

Also rename x to t for thread objects. I just like that better